src/requirements.txt
black~=22.0
flake8>=3.7.9, <5.0
flax~=0.7.4
ipython>=7.31.1, <8.0; python_version < '3.8'
ipython~=8.10; python_version >= '3.8'
isort~=5.0
jupyter~=1.0
jupyterlab_server>=2.11.1, <2.16.0
jupyterlab~=3.0, <3.6.0
kedro~=0.18.13
kedro-datasets[pandas.CSVDataSet, pandas.ParquetDataSet]~=1.0
kedro-telemetry~=0.2.0
nbstripout~=0.4
orbax-checkpoint~=0.4.1
optax~=0.1.7
pytest-cov~=3.0
pytest-mock>=1.7.1, <2.0
pytest~=7.2
tensorflow~=2.14.0